pyproject.toml
[project]
name = "vn_stock_advisor"
version = "0.4.1"
description = "vn_stock_advisor using Multiple AI Agent - powered by Google Gemini and CrewAI"
authors = [{ name = "Duong Anh Minh", email = "duonganhminhapple@gmail.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.117.0",
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.1.0",
    "vnstock>=3.2.4",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "litellm>=1.0.0",
]

[project.scripts]
vn_stock_advisor = "vn_stock_advisor.main:run"
run_crew = "vn_stock_advisor.main:run"
train = "vn_stock_advisor.main:train"
replay = "vn_stock_advisor.main:replay"
test = "vn_stock_advisor.main:test"
api_server = "vn_stock_advisor.api:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
